Fixed a bug with Podcasts. The problem was that for music files the playlist names are used to access the content of a playlist, where as for podcasts not the name but a reference itself was used. Thus "of user playlist plist" was working for music but not for podcasts.
Added support for audiobooks. All m4b files are automatically renamed to m4a.

Added support for m3u playlists. They are written with the ISO Latin 1 encoding, so also special characters (like German Umlaute äöü) should be handled by most players.
Currently already existing playlists on the player are not removed.

Fixed a bug when iTMW is configured with a negative limit to keep a certain amount of MB free. The older implementation checked for currently available free space and reduced it by the amount of space to be kept free. This works for the first sync, if the player is empty. In a second sync the so calculated total size limit is very small, and with only few (or even none) files this limit is reached. So iTMW collected just some few files (or even none) until that limit was reached and removed all other files from the music folder. So afterwards there were only very few or even no files on the player. Now /usr/bin/du is used to calculate the size currently in use by the music folder. It is added to the total size limit.

The calculation of the totalsize, the sum of all files to be transfered now is enhanced by rounding up to full block sizes. Probably the current assumed block size of 512 byte is not quite right for most players using FAT. It needs to be checked, if 1, 2 or 4 kB is a better block size assumption, probably it can be determined dynamically.

The second podcast folder level now gets the the podcast name. Similar for audiobooks.

For podcasts and audiobooks two m3u playlists Podcasts and Audiobooks are generated, named after to playlists, so their names are actually localized.

For music an overall m3u playlist Music is generated as well as a playlist per iTunes playlist to be synchronized. 
   
Replaced /bin/sleep by applescript command delay.

Podcast playlists have not been cleared before they were written.

On first creation of the default file, the setting for "SUCheckAtStartup" was missing and lead to an error message.