Added some try-blocks inside the getsongs function.
It happened that during a sync iTunes downloaded new podcast episodes and removed the one that was at that moment in use by the getsongs function. 

To prevent inconsistent global lists, I try to copy the important information to the end of the lists at once only at the end of blocks, where I don't expect errors any more because of removed tracks.

I didn't notice a degrade. However, it's hard to prove that my change is working as expected, as the failure I was facing is kind of a race condition. Even if it works as expected, it might still happen that later code parts have troubles, if a file is finally removed before or during the copying.