There was a problem with iso88959-1 encoding: Although the the entries of m3u playlist files were stripped to only contain valid iso88959-1 characters, still filenames could contain other utf-8 characters. Thus a playlist entry could refere to a file, which was not existing. Now file names (which are always encoded in utf-8) are also stripped down to the same character set as defined by iso88959-1 if necessary, to guarantee that playlist entries refere to proper files names.
I also improved the user feedback: Now the progress bar shows the progress while collecting the tracks from iTunes and while cleaning up the memory card. This means the progress bar is running three times now. Maybe each of this three steps should only run for a third of the whole progress bar, to not confuse people with three runs of that one progress bar.